updating logic to skip the restoreBuf region correctly if merged with neighbor memory regions.
The current checkpoint logic fails if the restoreBuf region is merged with
neighbor memory regions as the area's size is not exactly as restoreBufLen.
In some cases, restoreBuf may not even detected correctly if the area doesn't
starts with RestoreBufAddr. Unmapping and mapping back resolves the issue as
the region doesn't contain any data to lose and skipped from processing.